+++ src/Application/Query/Dashboard/DashboardMeetingViewQueryHandler.php
@@ -10,6 +10,7 @@
 
 namespace Proximum\Vimeet\Application\Query\Dashboard;
 
+use Proximum\Vimeet\Application\Query\Dashboard\View\DashboardRequestView;
 use Proximum\Vimeet\Application\View\Dashboard\DashboardMeetingView;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
@@ -61,9 +62,7 @@ class DashboardMeetingViewQueryHandler
         $meetingCreatedByPlanner = $this->meetingRepository
             ->countCreatedByEventAndType($query->event, Meeting::CREATED_BY_PLANNER);
 
-        $approvedRequest = $this->requestRepository->countApprovedByEvent($query->event);
-        $pendingRequest = $this->requestRepository->countPendingByEvent($query->event);
-        $refusedRequest = $this->requestRepository->countRefusedByEvent($query->event);
+        $dashboardRequestViews = $this->requestRepository->getDashboardRequestViewsByEvent($query->event);
 
         return new DashboardMeetingView(
             $allMeetings,
@@ -71,9 +70,118 @@ class DashboardMeetingViewQueryHandler
             $meetingCreatedByParticipant,
             $meetingCreatedByPlanner,
             $meetingCreatedUpstreamByAdmin,
-            $approvedRequest,
-            $pendingRequest,
-            $refusedRequest
+            $this->countApprovedRequests($dashboardRequestViews),
+            $this->countPendingRequests($dashboardRequestViews),
+            $this->countRefusedRequests($dashboardRequestViews),
+            $this->requestsByType($dashboardRequestViews),
+            $this->requestsByTypeAndState($dashboardRequestViews)
         );
+    }
+
+    /**
+     * @param DashboardRequestView[] $dashboardRequestViews
+     *
+     * @return int
+     */
+    private function countApprovedRequests(array &$dashboardRequestViews): int
+    {
+        $approved = 0;
+
+        foreach ($dashboardRequestViews as $dashboardRequestView) {
+            if ($dashboardRequestView->isApproved()) {
+                ++$approved;
+            }
+        }
+
+        return $approved;
+    }
+
+    /**
+     * @param DashboardRequestView[] $dashboardRequestViews
+     *
+     * @return int
+     */
+    private function countPendingRequests(array &$dashboardRequestViews): int
+    {
+        $pending = 0;
+
+        foreach ($dashboardRequestViews as $dashboardRequestView) {
+            if ($dashboardRequestView->isPending()) {
+                ++$pending;
+            }
+        }
+
+        return $pending;
+    }
+
+    /**
+     * @param DashboardRequestView[] $dashboardRequestViews
+     *
+     * @return int
+     */
+    private function countRefusedRequests(array &$dashboardRequestViews): int
+    {
+        $refused = 0;
+
+        foreach ($dashboardRequestViews as $dashboardRequestView) {
+            if ($dashboardRequestView->isRefused()) {
+                ++$refused;
+            }
+        }
+
+        return $refused;
+    }
+
+    /**
+     * @param DashboardRequestView[] $dashboardRequestViews
+     *
+     * @return array
+     */
+    private function requestsByType(array &$dashboardRequestViews): array
+    {
+        $requestsByType = [];
+
+        foreach ($dashboardRequestViews as $dashboardRequestView) {
+            $typeId = $dashboardRequestView->getFromTypeId();
+
+            if (!isset($requestsByType[$typeId])) {
+                $requestsByType[$typeId] = 1;
+            } else {
+                ++$requestsByType[$typeId];
+            }
+        }
+
+        return $requestsByType;
+    }
+
+    /**
+     * @param DashboardRequestView[] $dashboardRequestViews
+     *
+     * @return array
+     */
+    private function requestsByTypeAndState(array &$dashboardRequestViews): array
+    {
+        $requestsByTypeAndState = [];
+
+        foreach ($dashboardRequestViews as $dashboardRequestView) {
+            $typeId = $dashboardRequestView->getFromTypeId();
+            $state = $dashboardRequestView->getState();
+
+            if (!isset($requestsByTypeAndState[$typeId][$state])) {
+                $requestsByTypeAndState[$typeId][$state] = 1;
+            } else {
+                ++$requestsByTypeAndState[$typeId][$state];
+            }
+
+            if ($dashboardRequestView->isPlanned()) {
+                if (!isset($requestsByTypeAndState[$typeId][Meeting\Request::STATE_PLANNED])) {
+                    $requestsByTypeAndState[$typeId][Meeting\Request::STATE_PLANNED] = 1;
+                } else {
+                    ++$requestsByTypeAndState[$typeId][Meeting\Request::STATE_PLANNED];
+                }
+            }
+        }
+
+        return $requestsByTypeAndState;
     }
 }