+++ src/Application/View/Dashboard/DashboardView.php
@@ -10,6 +10,9 @@
 
 namespace Proximum\Vimeet\Application\View\Dashboard;
 
+use Proximum\Vimeet\Application\Query\Dashboard\View\DashboardContactView;
+use Proximum\Vimeet\Application\Query\Dashboard\View\DashboardEntranceScanView;
+
 class DashboardView
 {
     /** @var DashboardTransactionView */
@@ -21,18 +24,23 @@ class DashboardView
     /** @var DashboardMeetingView */
     public $dashboardMeetingView;
 
-    /**
-     * @param DashboardTransactionView $transactionView
-     * @param DashboardSheetView       $dashboardSheetView
-     * @param DashboardMeetingView     $dashboardMeetingView
-     */
+    /** @var DashboardContactView */
+    public $dashboardContactView;
+
+    /** @var DashboardEntranceScanView */
+    public $dashboardEntranceScanView;
+
     public function __construct(
         DashboardTransactionView $transactionView,
         DashboardSheetView $dashboardSheetView,
-        DashboardMeetingView $dashboardMeetingView
+        DashboardMeetingView $dashboardMeetingView,
+        DashboardContactView $dashboardContactView,
+        DashboardEntranceScanView $dashboardEntranceScanView
     ) {
         $this->transactionView    = $transactionView;
         $this->dashboardSheetView = $dashboardSheetView;
         $this->dashboardMeetingView = $dashboardMeetingView;
+        $this->dashboardContactView = $dashboardContactView;
+        $this->dashboardEntranceScanView = $dashboardEntranceScanView;
     }
 }