+++ src/Domain/Repository/MeetingRepositoryInterface.php
@@ -10,6 +10,7 @@
 
 namespace Proximum\Vimeet\Domain\Repository;
 
+use Proximum\Vimeet\Application\Query\Dashboard\View\DashboardMeetingContactEvaluationView;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\MeetingSlot;
@@ -295,4 +296,11 @@ interface MeetingRepositoryInterface
      * @return bool
      */
     public function hasAtLeastOneMeeting(array $sheets, array $otherSheets): bool;
+
+    /**
+     * @param Event $event
+     *
+     * @return DashboardMeetingContactEvaluationView[]
+     */
+    public function getDashboardMeetingContactEvaluationViews(Event $event): array;
 }