+++ src/Infrastructure/Repository/Meeting/RequestRepository.php
@@ -13,6 +13,7 @@ namespace Proximum\Vimeet\Infrastructure\Repository\Meeting;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\QueryBuilder;
 use Proximum\Vimeet\Application\Components\Paginator\Paginator;
+use Proximum\Vimeet\Application\Query\Dashboard\View\DashboardRequestView;
 use Proximum\Vimeet\Application\Query\MultipleSheets\Request\FilterRequestView;
 use Proximum\Vimeet\Application\View\Agenda\Slot\AvailableSlotView;
 use Proximum\Vimeet\Domain\Model\Event;
@@ -1115,5 +1116,20 @@ class RequestRepository implements RequestRepositoryInterface
         ;
 
         return null !== $queryBuilder->getQuery()->getOneOrNullResult();
+    }
+
+    public function getDashboardRequestViewsByEvent(Event $event): array
+    {
+        return $this
+            ->entityManager
+            ->createQueryBuilder()
+            ->select(sprintf('NEW %s(fromType.id, request.state, meeting.id)', DashboardRequestView::class))
+            ->from(Request::class, 'request')
+            ->join('request.from', 'fromSheet', 'WITH', 'request.event = :event AND request.disabled = false')
+            ->join('fromSheet.type', 'fromType')
+            ->leftJoin('request.meeting', 'meeting')
+            ->setParameter('event', $event)
+            ->getQuery()
+            ->getResult();
     }
 }