+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-1998 - Dashboard with contacts indicators
+- VIMEET-1997 - Dashboard with access control indicators
 
 ## [2.13.0] - 2019-10-09
 ### Added